+++ packages/react-core/src/components/Alert/AlertActionCloseButton.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Button, ButtonVariant, ButtonProps } from '../Button';
 import TimesIcon from '@patternfly/react-icons/dist/js/icons/times-icon';
-import { AlertContext } from '..';
+import { AlertContext } from './AlertContext';
 
 interface AlertActionCloseButtonProps extends ButtonProps {
   /** Additional classes added to the AlertActionCloseButton */
