+++ packages/react-core/src/components/Alert/AlertContext.ts
@@ -1,0 +1,8 @@
+import * as React from 'react';
+
+interface AlertContext {
+  title: React.ReactNode;
+  variantLabel?: string;
+}
+
+export const AlertContext = React.createContext<AlertContext>(null);
