+++ packages/react-core/src/components/ApplicationLauncher/ApplicationLauncherContext.ts
@@ -1,0 +1,5 @@
+import * as React from 'react';
+
+export const ApplicationLauncherContext = React.createContext({
+  onFavorite: (_itemId: string, _isFavorite: boolean) => {}
+});
