+++ packages/react-core/src/components/ApplicationLauncher/ApplicationLauncherItemContext.ts
@@ -1,0 +1,3 @@
+import * as React from 'react';
+
+export const ApplicationLauncherItemContext = React.createContext({ isExternal: false, icon: null });
